Modularize main, add user-agent, enhance logging
- Moved specific functions into separate modules for better organization: error_handling, logging_utils, user_input, user_interface, scraping_utils, session_utils
- Added a user-agent to HTTP requests for better web scraping behavior
- Enhanced logging by adding logging.info statements for "Creating a session..."
- Updated .gitignore
